pyproject.toml
[project]
name = "flaskapp"
version = "0.1.0"
description = "Demonstration Flask application with user authentication and session management."
readme = "README.md"
authors = [
    { name = "Eric Busboom", email = "eric@busboom.org" }
]
requires-python = ">=3.13"
dependencies = [
    "bootstrap-flask>=2.5.0",
    "flask>=3.1.1",
    "flask-login>=0.6.3",
    "flask-migrate>=4.1.0",
    "flask-session>=0.8.0",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "font-awesome-flask>=0.1.4",
    "gevent>=25.5.1",
    "gunicorn>=23.0.0",
    "jinja2>=3.1.6",
    "pip>=25.1.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "python-slugify>=8.0.4",
    "requests>=2.32.4",
    "tenacity>=9.1.2",
]

[project.scripts]
flaskapp = "flaskapp:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
[tool.hatch.build.targets.wheel]
packages = ["flaskapp"]
